requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.55.0
tiktoken>=0.8.0
chromadb>=0.5.0
tavily-python>=0.5.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
